removed obsolate functions from utilities_functions.py and moved other functions to tools.py
