+++ brahmap/utilities/__init__.py
@@ -3,46 +3,26 @@ This module contains the utilities functions strictly related to the computation
 
 """
 
-from .utilities_functions import (
-    is_sorted,
-    bash_colors,
-    filter_warnings,
-    profile_run,
-    output_profile,
-    rescalepixels,
-    angles_gen,
-    pairs_gen,
-    checking_output,
-    noise_val,
-    subscan_resize,
-    system_setup,
-)
-
 from .tools import (
+    bash_colors,
+    modify_numpy_context,
     TypeChangeWarning,
     LowerTypeCastWarning,
-    modify_numpy_context,
+    filter_warnings,
     ShapeError,
+    profile_run,
+    output_profile,
 )
 
 
 __all__ = [
-    # utilities_functions.py
-    "is_sorted",
-    "bash_colors",
-    "filter_warnings",
-    "profile_run",
-    "output_profile",
-    "rescalepixels",
-    "angles_gen",
-    "pairs_gen",
-    "checking_output",
-    "noise_val",
-    "subscan_resize",
-    "system_setup",
     # tools.py
+    "bash_colors",
+    "modify_numpy_context",
     "TypeChangeWarning",
     "LowerTypeCastWarning",
-    "modify_numpy_context",
+    "filter_warnings",
     "ShapeError",
+    "profile_run",
+    "output_profile",
 ]
