+++ brahmap/utilities/tools.py
@@ -1,5 +1,43 @@
+import warnings
 import numpy as np
 from ..math import parallel_norm
+
+
+class bash_colors:
+    """
+    This class contains the necessary definitions to print to bash
+    screen with colors. Sometimes it can be useful...
+    """
+
+    HEADER = "\033[95m"
+    OKBLUE = "\033[94m"
+    OKGREEN = "\033[92m"
+    WARNING = "\033[93m"
+    FAIL = "\033[91m"
+    ENDC = "\033[0m"
+    BOLD = "\033[1m"
+    UNDERLINE = "\033[4m"
+
+    def header(self, string):
+        return self.HEADER + str(string) + self.ENDC
+
+    def blue(self, string):
+        return self.OKBLUE + str(string) + self.ENDC
+
+    def green(self, string):
+        return self.OKGREEN + str(string) + self.ENDC
+
+    def warning(self, string):
+        return self.WARNING + str(string) + self.ENDC
+
+    def fail(self, string):
+        return self.FAIL + str(string) + self.ENDC
+
+    def bold(self, string):
+        return self.BOLD + str(string) + self.ENDC
+
+    def underline(self, string):
+        return self.UNDERLINE + str(string) + self.ENDC
 
 
 class modify_numpy_context(object):
@@ -30,6 +68,16 @@ class LowerTypeCastWarning(Warning):
         return repr(self.message)
 
 
+def filter_warnings(wfilter):
+    """
+    wfilter: {string}
+    - "ignore": never print matching warnings;
+    - "always": always print matching warnings
+
+    """
+    warnings.simplefilter(wfilter)
+
+
 class ShapeError(Exception):
     """
     Exception class for handling shape mismatch errors.
@@ -45,3 +93,34 @@ class ShapeError(Exception):
 
     def __str__(self):
         return repr(self.value)
+
+
+def profile_run():
+    """
+    Profile the execution with :mod:`cProfile`
+    """
+    import cProfile
+
+    pr = cProfile.Profile()
+    return pr
+
+
+def output_profile(pr):
+    """
+    Output of the profiling with :func:`profile_run`.
+
+    **Parameter**
+
+    - ``pr``:
+        instance returned by :func:`profile_run`
+
+    """
+    import pstats
+    import io
+
+    s = io.StringIO()
+    sortby = "cumulative"
+    ps = pstats.Stats(pr, stream=s).sort_stats(sortby)
+    ps.print_stats()
+    print(s.getvalue())
+    pass
